+++ src/main/java/kr/ac/kumoh/d138/JobForeigner/job/dto/request/JobPostRequestDto.java
@@ -1,5 +1,6 @@
-package kr.ac.kumoh.d138.JobForeigner.job.dto.company.request;
+package kr.ac.kumoh.d138.JobForeigner.job.dto.request;
 
+import jakarta.validation.constraints.NotNull;
 import kr.ac.kumoh.d138.JobForeigner.job.domain.JobPostStatus;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -7,17 +8,28 @@ import lombok.NoArgsConstructor;
 import java.time.LocalDateTime;
 
 @Getter
-@NoArgsConstructor
-public class JobTempPostRequestDto {
+@NoArgsConstructor 
+public class JobPostRequestDto {
+    @NotNull
     private Long id;
+    @NotNull
     private String title;
+    @NotNull
     private String description;
+    @NotNull
     private String location;
+    @NotNull
     private String employmentType;
+    @NotNull
     private String salary;
+    @NotNull
     private String career;
+    @NotNull
     private JobPostStatus published;
+    @NotNull
     private String grade;
+    @NotNull
     private LocalDateTime expiryAt;
+    @NotNull
     private Long companyId;
 }
