+++ src/main/java/kr/ac/kumoh/d138/JobForeigner/job/dto/request/JobTempPostRequestDto.java
@@ -1,6 +1,5 @@
-package kr.ac.kumoh.d138.JobForeigner.job.dto.company.request;
+package kr.ac.kumoh.d138.JobForeigner.job.dto.request;
 
-import jakarta.validation.constraints.NotNull;
 import kr.ac.kumoh.d138.JobForeigner.job.domain.JobPostStatus;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -8,28 +7,17 @@ import lombok.NoArgsConstructor;
 import java.time.LocalDateTime;
 
 @Getter
-@NoArgsConstructor 
-public class JobPostRequestDto {
-    @NotNull
+@NoArgsConstructor
+public class JobTempPostRequestDto {
     private Long id;
-    @NotNull
     private String title;
-    @NotNull
     private String description;
-    @NotNull
     private String location;
-    @NotNull
     private String employmentType;
-    @NotNull
     private String salary;
-    @NotNull
     private String career;
-    @NotNull
     private JobPostStatus published;
-    @NotNull
     private String grade;
-    @NotNull
     private LocalDateTime expiryAt;
-    @NotNull
     private Long companyId;
 }
