+++ src/main/java/kr/ac/kumoh/d138/JobForeigner/job/dto/response/JobPostDetailResponseDto.java
@@ -1,4 +1,4 @@
-package kr.ac.kumoh.d138.JobForeigner.job.dto.company.response;
+package kr.ac.kumoh.d138.JobForeigner.job.dto.response;
 
 import kr.ac.kumoh.d138.JobForeigner.job.domain.JobPost;
 import kr.ac.kumoh.d138.JobForeigner.job.domain.JobPostStatus;
@@ -8,7 +8,7 @@ import lombok.Setter;
 
 @Getter
 @Setter
-public class UpdateJobPostResponseDto {
+public class JobPostDetailResponseDto {
     private Long id;
     private String title;
     private String description;
@@ -18,9 +18,12 @@ public class UpdateJobPostResponseDto {
     private String career;
     private JobPostStatus published;
     private String grade;
+    private Boolean isScrapped;
 
     @Builder
-    public UpdateJobPostResponseDto(Long id, String title, String description, String location, String employmentType, String salary, String career, JobPostStatus published, String grade) {
+    public JobPostDetailResponseDto(Long id, String title, String description, String location,
+                                    String employmentType, String salary, String career,
+                                    JobPostStatus published, String grade, Boolean isScrapped) {
         this.id = id;
         this.title = title;
         this.description = description;
@@ -30,10 +33,11 @@ public class UpdateJobPostResponseDto {
         this.career = career;
         this.published = published;
         this.grade = grade;
+        this.isScrapped = isScrapped;
     }
 
-    public static UpdateJobPostResponseDto fromEntity(JobPost jobPost) {
-        return UpdateJobPostResponseDto.builder()
+    public static JobPostDetailResponseDto fromEntity(JobPost jobPost, boolean isScrapped) {
+        return JobPostDetailResponseDto.builder()
                 .id(jobPost.getId())
                 .title(jobPost.getTitle())
                 .description(jobPost.getDescription())
@@ -43,6 +47,7 @@ public class UpdateJobPostResponseDto {
                 .career(jobPost.getCareer())
                 .published(jobPost.getPublished())
                 .grade(jobPost.getGrade())
+                .isScrapped(isScrapped)
                 .build();
     }
 }
