+++ src/main/java/kr/ac/kumoh/d138/JobForeigner/job/dto/response/JobPostResponseDto.java
@@ -1,4 +1,4 @@
-package kr.ac.kumoh.d138.JobForeigner.job.dto.company.response;
+package kr.ac.kumoh.d138.JobForeigner.job.dto.response;
 
 import kr.ac.kumoh.d138.JobForeigner.job.domain.JobPost;
 import kr.ac.kumoh.d138.JobForeigner.job.domain.JobPostStatus;
@@ -8,46 +8,43 @@ import lombok.Setter;
 
 @Getter
 @Setter
-public class JobPostDetailResponseDto {
+public class JobPostResponseDto {
     private Long id;
     private String title;
     private String description;
     private String location;
-    private String employmentType;
+    private String employment_type;
     private String salary;
     private String career;
     private JobPostStatus published;
     private String grade;
-    private Boolean isScrapped;
 
     @Builder
-    public JobPostDetailResponseDto(Long id, String title, String description, String location,
-                                    String employmentType, String salary, String career,
-                                    JobPostStatus published, String grade, Boolean isScrapped) {
+    public JobPostResponseDto(Long id, String title, String description, String location,
+                              String employment_type, String salary, String career,
+                              JobPostStatus published, String grade) {
         this.id = id;
         this.title = title;
         this.description = description;
         this.location = location;
-        this.employmentType = employmentType;
+        this.employment_type = employment_type;
         this.salary = salary;
         this.career = career;
         this.published = published;
         this.grade = grade;
-        this.isScrapped = isScrapped;
     }
 
-    public static JobPostDetailResponseDto fromEntity(JobPost jobPost, boolean isScrapped) {
-        return JobPostDetailResponseDto.builder()
+    public static JobPostResponseDto fromEntity(JobPost jobPost) {
+        return JobPostResponseDto.builder()
                 .id(jobPost.getId())
                 .title(jobPost.getTitle())
                 .description(jobPost.getDescription())
                 .location(jobPost.getLocation())
-                .employmentType(jobPost.getEmploymentType())
+                .employment_type(jobPost.getEmploymentType())
                 .salary(jobPost.getSalary())
                 .career(jobPost.getCareer())
                 .published(jobPost.getPublished())
                 .grade(jobPost.getGrade())
-                .isScrapped(isScrapped)
                 .build();
     }
 }
