+++ src/main/java/kr/ac/kumoh/d138/JobForeigner/job/dto/response/UpdateJobPostResponseDto.java
@@ -1,4 +1,4 @@
-package kr.ac.kumoh.d138.JobForeigner.job.dto.company.response;
+package kr.ac.kumoh.d138.JobForeigner.job.dto.response;
 
 import kr.ac.kumoh.d138.JobForeigner.job.domain.JobPost;
 import kr.ac.kumoh.d138.JobForeigner.job.domain.JobPostStatus;
@@ -8,39 +8,37 @@ import lombok.Setter;
 
 @Getter
 @Setter
-public class JobPostResponseDto {
+public class UpdateJobPostResponseDto {
     private Long id;
     private String title;
     private String description;
     private String location;
-    private String employment_type;
+    private String employmentType;
     private String salary;
     private String career;
     private JobPostStatus published;
     private String grade;
 
     @Builder
-    public JobPostResponseDto(Long id, String title, String description, String location,
-                              String employment_type, String salary, String career,
-                              JobPostStatus published, String grade) {
+    public UpdateJobPostResponseDto(Long id, String title, String description, String location, String employmentType, String salary, String career, JobPostStatus published, String grade) {
         this.id = id;
         this.title = title;
         this.description = description;
         this.location = location;
-        this.employment_type = employment_type;
+        this.employmentType = employmentType;
         this.salary = salary;
         this.career = career;
         this.published = published;
         this.grade = grade;
     }
 
-    public static JobPostResponseDto fromEntity(JobPost jobPost) {
-        return JobPostResponseDto.builder()
+    public static UpdateJobPostResponseDto fromEntity(JobPost jobPost) {
+        return UpdateJobPostResponseDto.builder()
                 .id(jobPost.getId())
                 .title(jobPost.getTitle())
                 .description(jobPost.getDescription())
                 .location(jobPost.getLocation())
-                .employment_type(jobPost.getEmploymentType())
+                .employmentType(jobPost.getEmploymentType())
                 .salary(jobPost.getSalary())
                 .career(jobPost.getCareer())
                 .published(jobPost.getPublished())
